# [cmis] combine cmis eeprom i2c read